+++ src/test/java/nextstep/utils/DataLoaderBootstrap.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.utils;
 
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.annotation.Profile;
